# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "OS Backend API service",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "https://test-api.seedao.tech",
    "basePath": "/v1",
    "paths": {
...
        "/public_data/notion/user/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PublicData"
                ],
                "summary": "NotionUser returns the detail of a notion user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Reply"
                        }
                    }
                }
            }
        },
...
        "/user/me": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user detail",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Reply"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sdk.SeepassResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user info",
                "parameters": [
                    {
                        "description": "request json body",
                        "name": "JsonBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Reply"
                        }
                    }
                }
            }
        },
...
        "/user/users": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Query multiple users by wallet array on batch",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "wallets",
                        "name": "wallets",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Reply"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/user.UserModelWithSomeSeepassData"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Reply": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.UserAssetRecord": {
            "type": "object",
            "properties": {
                "asset_name": {
                    "description": "asset name",
                    "type": "string"
                },
                "create_ts": {
                    "type": "integer"
                },
                "dealt_amount": {
                    "description": "amount of asset that already dealt",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "processing_amount": {
                    "description": "amount of asset that still need confirmation",
                    "type": "number"
                },
                "update_ts": {
                    "type": "integer"
                },
                "user_wallet": {
                    "type": "string"
                }
            }
        },
...
        "sdk.SeepassResponse": {
            "type": "object",
            "properties": {
                "avatar": {},
                "bio": {},
                "email": {},
                "level": {
                    "type": "object",
                    "properties": {
                        "current_lv": {
                            "type": "string"
                        },
                        "next_lv": {
                            "type": "string"
                        },
                        "scr_to_next_lv": {
                            "type": "string"
                        },
                        "upgrade_percent": {
                            "type": "string"
                        }
                    }
                },
                "nickname": {},
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sbt": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "collection_name": {
                                "type": "string"
                            },
                            "contract_addr": {
                                "type": "string"
                            },
                            "contract_type": {
                                "type": "string"
                            },
                            "image_uri": {
                                "type": "string"
                            },
                            "metadata": {},
                            "name": {
                                "type": "string"
                            },
                            "symbol": {
                                "type": "string"
                            },
                            "token_amount": {
                                "type": "string"
                            },
                            "token_id": {
                                "type": "string"
                            }
                        }
                    }
                },
                "scr": {
                    "type": "object",
                    "properties": {
                        "amount": {
                            "type": "string"
                        },
                        "contract_addr": {
                            "type": "string"
                        }
                    }
                },
                "seed": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "contract_addr": {
                                "type": "string"
                            },
                            "contract_type": {
                                "type": "string"
                            },
                            "image_uri": {
                                "type": "string"
                            },
                            "token_amount": {
                                "type": "string"
                            },
                            "token_id": {
                                "type": "string"
                            }
                        }
                    }
                },
                "sns": {
                    "type": "string"
                },
                "social_accounts": {
                    "type": "array",
                    "items": {}
                },
                "wallet": {
                    "type": "string"
                }
            }
        },
...
        "user.UpdateReq": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "discord_profile": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "github_profile": {
                    "type": "string"
                },
                "google_profile": {
                    "type": "string"
                },
                "mirror": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "twitter_profile": {
                    "type": "string"
                },
                "wechat": {
                    "type": "string"
                }
            }
        },
        "user.UserModelWithSomeSeepassData": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserAssetRecord"
                    }
                },
                "avatar": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "create_ts": {
                    "type": "integer"
                },
                "discord_profile": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "github_profile": {
                    "type": "string"
                },
                "google_profile": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mirror": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sp": {
                    "$ref": "#/definitions/sdk.SeepassResponse"
                },
                "twitter_profile": {
                    "type": "string"
                },
                "update_ts": {
                    "type": "integer"
                },
                "wallet": {
                    "type": "string"
                },
                "wechat": {
                    "type": "string"
                }
            }
        }
    }
}